Implemented the first function to provide new helpers to the framework
I added the `rods` concept to the framework. Their aim is provide
helper functions which allow the implement `engine` sequence elements or
be helpful to simplify the implementation of them.

I added two base functions in `rods/browser` which allow to inject
javascript browser code written into a function, therefore the browser
script can be written in an editor and execute rather than in a string,
as webdriver expects to receive as a first argument of the methods
executeAsyncScript and executeScript.

I also provided the test for this base functions and added to
package.json the new dependencies required